requirements.txt
numpy>=1.21.2
matplotlib>=3.6.3
Pillow>=9.4.0